add feature_home module
