+++ .idea/compiler.xml
@@ -18,6 +18,7 @@
       <module name="Kizzy.feature_console_rpc" target="17" />
       <module name="Kizzy.feature_crash_handler" target="17" />
       <module name="Kizzy.feature_custom_rpc" target="17" />
+      <module name="Kizzy.feature_home" target="17" />
       <module name="Kizzy.feature_logs" target="17" />
       <module name="Kizzy.feature_media_rpc" target="17" />
       <module name="Kizzy.feature_profile" target="17" />
