+++ .idea/gradle.xml
@@ -24,6 +24,7 @@
             <option value="$PROJECT_DIR$/feature_console_rpc" />
             <option value="$PROJECT_DIR$/feature_crash_handler" />
             <option value="$PROJECT_DIR$/feature_custom_rpc" />
+            <option value="$PROJECT_DIR$/feature_home" />
             <option value="$PROJECT_DIR$/feature_logs" />
             <option value="$PROJECT_DIR$/feature_media_rpc" />
             <option value="$PROJECT_DIR$/feature_profile" />
