+++ feature_home/src/main/res/drawable/ic_console_games.xml
@@ -1,0 +1,25 @@
+<!--
+  ~ /*
+  ~ * ******************************************************************
+  ~  *  * Copyright (C) 2022
+  ~  *  * ic_console_games.xml is part of Kizzy
+  ~  *  *  and can not be copied and/or distributed without the express
+  ~  *  * permission of yzziK(Vaibhav)
+  ~  *  *****************************************************************
+  ~  */
+  ~
+  -->
+
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:width="48dp"
+    android:height="48dp"
+    android:viewportWidth="48"
+    android:viewportHeight="48"
+    android:tint="?attr/colorControlNormal">
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M9.45,40Q6.3,40 4.25,37.8Q2.2,35.6 2.1,32.65Q2.1,32.25 2.15,31.75Q2.2,31.25 2.3,30.85L6.5,14.05Q7.2,11.4 9.375,9.7Q11.55,8 14.25,8H33.75Q36.45,8 38.625,9.7Q40.8,11.4 41.5,14.05L45.7,30.85Q45.8,31.3 45.875,31.775Q45.95,32.25 45.95,32.7Q45.95,35.7 43.825,37.85Q41.7,40 38.55,40Q36.45,40 34.65,38.9Q32.85,37.8 31.95,35.9L30.55,33Q30.3,32.45 29.825,32.225Q29.35,32 28.75,32H19.25Q18.65,32 18.175,32.225Q17.7,32.45 17.45,33L16.05,35.9Q15.15,37.8 13.35,38.9Q11.55,40 9.45,40ZM9.6,37Q10.7,37 11.725,36.425Q12.75,35.85 13.4,34.6L14.8,31.75Q15.45,30.45 16.625,29.725Q17.8,29 19.25,29H28.75Q30.15,29 31.35,29.75Q32.55,30.5 33.25,31.75L34.65,34.6Q35.3,35.85 36.325,36.425Q37.35,37 38.45,37Q40.15,37 41.5,35.825Q42.85,34.65 42.9,32.75Q42.9,32.5 42.875,32.175Q42.85,31.85 42.75,31.55L38.55,14.8Q38.15,13.1 36.825,12.05Q35.5,11 33.75,11H14.25Q12.5,11 11.175,12.05Q9.85,13.1 9.45,14.8L5.25,31.55Q5.2,31.75 5.1,32.7Q5.1,34.6 6.5,35.8Q7.9,37 9.6,37ZM27,21.5Q26.4,21.5 25.95,21.05Q25.5,20.6 25.5,19.95Q25.5,19.35 25.95,18.925Q26.4,18.5 27.05,18.5Q27.65,18.5 28.075,18.925Q28.5,19.35 28.5,20Q28.5,20.6 28.075,21.05Q27.65,21.5 27,21.5ZM31,17.5Q30.4,17.5 29.95,17.05Q29.5,16.6 29.5,15.95Q29.5,15.35 29.95,14.925Q30.4,14.5 31.05,14.5Q31.65,14.5 32.075,14.925Q32.5,15.35 32.5,16Q32.5,16.6 32.075,17.05Q31.65,17.5 31,17.5ZM31,25.5Q30.4,25.5 29.95,25.05Q29.5,24.6 29.5,23.95Q29.5,23.35 29.95,22.925Q30.4,22.5 31.05,22.5Q31.65,22.5 32.075,22.925Q32.5,23.35 32.5,24Q32.5,24.6 32.075,25.05Q31.65,25.5 31,25.5ZM35,21.5Q34.4,21.5 33.95,21.05Q33.5,20.6 33.5,19.95Q33.5,19.35 33.95,18.925Q34.4,18.5 35.05,18.5Q35.65,18.5 36.075,18.925Q36.5,19.35 36.5,20Q36.5,20.6 36.075,21.05Q35.65,21.5 35,21.5ZM17,24.75Q16.5,24.75 16.125,24.375Q15.75,24 15.75,23.5V21.25H13.5Q13,21.25 12.625,20.875Q12.25,20.5 12.25,20Q12.25,19.5 12.625,19.125Q13,18.75 13.5,18.75H15.75V16.5Q15.75,16 16.125,15.625Q16.5,15.25 17,15.25Q17.5,15.25 17.875,15.625Q18.25,16 18.25,16.5V18.75H20.5Q21,18.75 21.375,19.125Q21.75,19.5 21.75,20Q21.75,20.5 21.375,20.875Q21,21.25 20.5,21.25H18.25V23.5Q18.25,24 17.875,24.375Q17.5,24.75 17,24.75Z"
+        tools:ignore="VectorPath" />
+</vector>
+
