+++ feature_home/src/main/res/drawable/ic_dev_rpc.xml
@@ -1,0 +1,11 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:width="48dp"
+    android:height="48dp"
+    android:viewportWidth="48"
+    android:viewportHeight="48">
+  <path
+      android:fillColor="@android:color/white"
+      android:pathData="M17.8,42L9,42q-1.2,0 -2.1,-0.9Q6,40.2 6,39v-8.8q2.2,-0.25 3.775,-1.725Q11.35,27 11.35,24.85t-1.575,-3.625Q8.2,19.75 6,19.5v-8.8q0,-1.2 0.9,-2.1 0.9,-0.9 2.1,-0.9h8.85q0.55,-2 1.975,-3.35T23.25,3q2,0 3.425,1.35Q28.1,5.7 28.65,7.7h8.65q1.2,0 2.1,0.9 0.9,0.9 0.9,2.1v8.65q2,0.55 3.275,2.075 1.275,1.525 1.275,3.525t-1.275,3.35Q42.3,29.65 40.3,30.2L40.3,39q0,1.2 -0.9,2.1 -0.9,0.9 -2.1,0.9h-8.8q-0.25,-2.4 -1.775,-3.875T23.15,36.65q-2.05,0 -3.575,1.475Q18.05,39.6 17.8,42ZM9,39h6.5q1.25,-3.05 3.5,-4.2 2.25,-1.15 4.15,-1.15 1.9,0 4.15,1.15 2.25,1.15 3.5,4.2h6.5L37.3,27.25h2.25q1,0 1.65,-0.65 0.65,-0.65 0.65,-1.65 0,-1 -0.65,-1.65 -0.65,-0.65 -1.65,-0.65L37.3,22.65L37.3,10.7L25.55,10.7L25.55,8.3q0,-1 -0.65,-1.65Q24.25,6 23.25,6q-1,0 -1.65,0.65 -0.65,0.65 -0.65,1.65v2.4L9,10.7v6.5q2.4,0.9 3.875,3t1.475,4.65q0,2.55 -1.475,4.65Q11.4,31.6 9,32.5ZM25.45,22.5Z"
+      tools:ignore="VectorPath" />
+</vector>
