+++ feature_home/src/main/res/drawable/ic_media_rpc.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="48dp"
+    android:height="48dp"
+    android:viewportWidth="48"
+    android:viewportHeight="48">
+    <path
+        android:fillColor="@android:color/white"
+        android:pathData="M7,8 L10.7,15.6H17.2L13.5,8H17.95L21.65,15.6H28.15L24.45,8H28.9L32.6,15.6H39.1L35.4,8H41Q42.2,8 43.1,8.9Q44,9.8 44,11V37Q44,38.2 43.1,39.1Q42.2,40 41,40H7Q5.8,40 4.9,39.1Q4,38.2 4,37V11Q4,9.8 4.9,8.9Q5.8,8 7,8ZM7,18.6V37Q7,37 7,37Q7,37 7,37H41Q41,37 41,37Q41,37 41,37V18.6ZM7,18.6V37Q7,37 7,37Q7,37 7,37Q7,37 7,37Q7,37 7,37Z"/>
+</vector>
