+++ feature_home/src/main/res/drawable/ic_rpc_placeholder.xml
@@ -1,0 +1,36 @@
+<!--
+  ~ /*
+  ~ * ******************************************************************
+  ~  *  * Copyright (C) 2022
+  ~  *  * ic_rpc_placeholder.xml is part of Kizzy
+  ~  *  *  and can not be copied and/or distributed without the express
+  ~  *  * permission of yzziK(Vaibhav)
+  ~  *  *****************************************************************
+  ~  */
+  ~
+  -->
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="48dp"
+    android:height="48dp"
+    android:viewportWidth="40"
+    android:viewportHeight="40">
+    <path
+        android:pathData="M21.226,30v-4.154h-3.755L17.471,30h3.755zM20.714,21.38c3.67,-1.337 5.718,-2.959 5.718,-6.145 0,-3.272 -2.36,-5.235 -6.088,-5.235 -2.589,0 -4.637,0.825 -6.344,2.048l0.484,3.3c1.621,-1.251 3.47,-2.162 5.49,-2.162 1.707,0 2.845,0.854 2.845,2.305 0,1.593 -1.08,2.475 -3.812,3.442l-0.882,0.342 0.342,4.523 2.076,-0.455 0.171,-1.963z"
+        android:fillColor="@android:color/white" />
+    <path
+        android:pathData="M6,1L34,1A5,5 0,0 1,39 6L39,34A5,5 0,0 1,34 39L6,39A5,5 0,0 1,1 34L1,6A5,5 0,0 1,6 1z"
+        android:strokeWidth="3"
+        android:strokeColor="@android:color/white"/>
+    <path
+        android:pathData="M7,7m-2,0a2,2 0,1 1,4 0a2,2 0,1 1,-4 0"
+        android:fillColor="@android:color/white" />
+    <path
+        android:pathData="M7,33m-2,0a2,2 0,1 1,4 0a2,2 0,1 1,-4 0"
+        android:fillColor="@android:color/white"/>
+    <path
+        android:pathData="M33,7m-2,0a2,2 0,1 1,4 0a2,2 0,1 1,-4 0"
+        android:fillColor="@android:color/white"/>
+    <path
+        android:pathData="M33,33m-2,0a2,2 0,1 1,4 0a2,2 0,1 1,-4 0"
+        android:fillColor="@android:color/white"/>
+</vector>
